+++ ke-legacy/main.js
@@ -19,6 +19,7 @@ service.vars.currency = 'KES';
 
 service.vars.topUpStart = env + '_start_top_up';
 service.vars.topUpEnd = env + '_end_top_up';
+service.vars.seed_germination_issues_table = env + '_seed_germination_issues';
 
 var notifyELK = require('../notifications/elk-notification/elkNotification');
 var transactionHistory = require('../transaction-history/transactionHistory');
@@ -29,6 +30,7 @@ var rosterAPI = require('../rw-legacy/lib/roster/api');
 var dukaClient = require('../duka-client/dukaClient');
 var isCreditOfficer = require('../duka-client/checkCreditOfficer');
 var warrantyExpiration = require('../warranty-expiration/warrantyExpiration');
+var seedGerminationIssues = require('../seed-germination-issues/seedGerminationIssues');
 
 var slackLogger = require('../slack-logger/index');
 var Log = require('../logger/elk/elk-logger');
@@ -1616,6 +1618,8 @@ global.main = function () {
     promptDigits('SplashMenu', {submitOnHash: true, maxDigits: 8, timeout: 5});
 };
 
+var langWithEnke = GetLang() ? 'en-ke' : 'sw';
+
 // load input handlers
 dukaLocator.registerDukaLocatorHandlers({lang: GetLang() ? 'en' : 'sw'});
 transactionHistory.registerHandlers();
@@ -1624,7 +1628,7 @@ justInTime.registerHandlers();
 groupRepaymentsModule.registerGroupRepaymentHandlers({lang: GetLang() ? 'en' : 'sw', main_menu: state.vars.main_menu, main_menu_handler: 'MainMenu'});
 dukaClient.registerInputHandlers(GetLang() ? 'en-ke' : 'sw', service.vars.duka_clients_table);
 warrantyExpiration.registerHandlers();
-
+seedGerminationIssues.registerInputHandlers(langWithEnke, service.vars.seed_germination_issues_table);
 
 function reduceClientSize(client) {
     var cloned = _.clone(client);
@@ -1722,8 +1726,10 @@ addInputHandler('NonClientMenu', function(input) {
     }
     else if(sessionMenu[input-1].option_name == 'locate_oaf_duka') {
         dukaLocator.startDukaLocator({lang: GetLang() ? 'en' : 'sw'});
-    }
-    else{
+    } else if(sessionMenu[input-1].option_name === 'report_seed_quality') {
+        //start the seed germination issues
+        seedGerminationIssues.start(langWithEnke);
+    } else{
         NonClientMenuText();
         promptDigits('NonClientMenu', {submitOnHash: true, maxDigits: 2, timeout: 5});
     }
@@ -1854,6 +1860,10 @@ addInputHandler('MainMenu', function(SplashMenu){
     }
     else if (sessionMenu[SplashMenu - 1].option_name == 'warranty_expiration') {
         warrantyExpiration.start(client.GlobalClientId, state.vars.lang);
+    }
+    else if(sessionMenu[SplashMenu - 1].option_name ===  'report_seed_quality') {
+        //start the seed germination issues
+        seedGerminationIssues.start(langWithEnke);
     }
     else{
         var arrayLength = client.BalanceHistory.length;